Move utils related tests to utils package
